modularize github parsing code
